Move the mirror's fixtures and slice-name helpers into the module (#81297)
* Move the mirror's fixtures and slice-name helpers into the module

* Build the QuestionSources fixture through the metadata store barrel

* Document why the metadata store is not a deletable mirror slice
