+++ frontend/src/metabase-lib/v1/metadata/utils/saved-questions.ts
@@ -79,7 +79,7 @@ export function convertSavedQuestionToVirtualTable(card: ConvertibleCard) {
     moderated_status: card.moderated_status,
     // we may not have permissions
     db_id: card.dataset_query?.database,
-    type: "question",
+    type: "question" as const,
     schema: getCollectionVirtualSchemaId(card.collection),
     schema_name: getCollectionVirtualSchemaName(card.collection),
   };
