+++ frontend/src/metabase/redux/store/entities.ts
@@ -10,19 +10,6 @@ import type {
   NormalizedTable,
 } from "metabase-types/api";
 
-// backend returns model = "card" instead of "question"
-export const entityTypeForModel = (model: string): string => {
-  if (model === "card" || model === "dataset" || model === "metric") {
-    return "questions";
-  }
-  return `${model}s`;
-};
-
-export const entityTypeForObject = (
-  object?: { model: string } | null,
-): string | undefined =>
-  object ? entityTypeForModel(object.model) : undefined;
-
 export interface EntitiesState {
   databases: Record<string, NormalizedDatabase>;
   schemas: Record<string, NormalizedSchema>;
